Update RoyaltiesRegistry.sol
Rather than do an ERC165 check on the contract for the v1 interface, simply make the call to getFeeRecipients/getFeeBps.  It ends up using less gas, and an unimplemeneted contract would fail anyways, resulting in the same behavior as before.

Further, it would automatically handle projects that implemented the spec but not the ERC165 check.